+++ lib/redmine_custom_workflows/patches/mailer_patch.rb
@@ -4,7 +4,7 @@
 # Redmine plugin for Custom Workflows
 #
 # Copyright © 2015-19 Anton Argirov
-# Copyright © 2019-21 Karel Pičman <karel.picman@kontron.com>
+# Copyright © 2019-22 Karel Pičman <karel.picman@kontron.com>
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
@@ -51,5 +51,9 @@ module RedmineCustomWorkflows
   end
 end
 
-RedmineExtensions::PatchManager.register_model_patch 'Mailer',
-  'RedmineCustomWorkflows::Patches::MailerPatch'
+# Apply the patch
+if Redmine::Plugin.installed?(:easy_extensions)
+  RedmineExtensions::PatchManager.register_model_patch 'Mailer', 'RedmineCustomWorkflows::Patches::MailerPatch'
+else
+  Mailer.prepend RedmineCustomWorkflows::Patches::MailerPatch
+end