+++ lib/redmine_custom_workflows/patches/project_patch.rb
@@ -4,7 +4,7 @@
 # Redmine plugin for Custom Workflows
 #
 # Copyright © 2015-19 Anton Argirov
-# Copyright © 2019-21 Karel Pičman <karel.picman@kontron.com>
+# Copyright © 2019-22 Karel Pičman <karel.picman@kontron.com>
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
@@ -24,7 +24,7 @@ module RedmineCustomWorkflows
   module Patches
     module ProjectPatch
 
-      def self.included(base)
+      def self.prepended(base)
         base.class_eval do
           has_and_belongs_to_many :custom_workflows
           safe_attributes :custom_workflow_ids, if:
@@ -75,6 +75,9 @@ module RedmineCustomWorkflows
   end
 end
 
-# Apply patch
-RedmineExtensions::PatchManager.register_model_patch 'Project',
-  'RedmineCustomWorkflows::Patches::ProjectPatch'
+# Apply the patch
+if Redmine::Plugin.installed?(:easy_extensions)
+  RedmineExtensions::PatchManager.register_model_patch 'Project', 'RedmineCustomWorkflows::Patches::ProjectPatch'
+else
+  Project.prepend RedmineCustomWorkflows::Patches::ProjectPatch
+end