+++ lib/redmine_custom_workflows/patches/projects_helper_patch.rb
@@ -4,7 +4,7 @@
 # Redmine plugin for Custom Workflows
 #
 # Copyright © 2015-19 Anton Argirov
-# Copyright © 2019-21 Karel Pičman <karel.picman@kontron.com>
+# Copyright © 2019-22 Karel Pičman <karel.picman@kontron.com>
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
@@ -35,6 +35,7 @@ module RedmineCustomWorkflows
   end
 end
 
+# Apply the patch
 if Redmine::Plugin.installed?(:easy_extensions)
   RedmineExtensions::PatchManager.register_helper_patch 'ProjectsHelper',
     'RedmineCustomWorkflows::Patches::ProjectsHelperPatch', prepend: true