+++ lib/redmine_custom_workflows/patches/wiki_page_patch.rb
@@ -4,7 +4,7 @@
 # Redmine plugin for Custom Workflows
 #
 # Copyright © 2015-19 Anton Argirov
-# Copyright © 2019-21 Karel Pičman <karel.picman@kontron.com>
+# Copyright © 2019-22 Karel Pičman <karel.picman@kontron.com>
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
@@ -24,7 +24,7 @@ module RedmineCustomWorkflows
   module Patches
     module WikiPagePatch
 
-      def self.included(base)
+      def self.prepended(base)
         base.class_eval do
           def self.attachments_callback(event, page, attachment)
             page.instance_variable_set :@page, page
@@ -43,6 +43,9 @@ module RedmineCustomWorkflows
   end
 end
 
-# Apply patch
-RedmineExtensions::PatchManager.register_model_patch 'WikiPage',
-  'RedmineCustomWorkflows::Patches::WikiPagePatch'
+# Apply the patch
+if Redmine::Plugin.installed?(:easy_extensions)
+  RedmineExtensions::PatchManager.register_model_patch 'WikiPage', 'RedmineCustomWorkflows::Patches::WikiPagePatch'
+else
+  WikiPage.prepend RedmineCustomWorkflows::Patches::WikiPagePatch
+end